rename: 야놀자 사용자 dummy 패키지로 이동 &  API Prefix "/dummy"로 변경
