+++ src/main/java/site/goldenticket/dummy/yauser/controller/YaUserController.java
@@ -1,13 +1,14 @@
-package site.goldenticket.domain.yauser.controller;
+package site.goldenticket.dummy.yauser.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import site.goldenticket.domain.yauser.service.YaUserService;
+import site.goldenticket.dummy.yauser.service.YaUserService;
 
 @RestController
-@RequestMapping("/yausers")
+@RequestMapping("/dummy")
 @RequiredArgsConstructor
 public class YaUserController {
+
     private final YaUserService yaUserService;
 }
