+++ src/main/java/site/goldenticket/dummy/yauser/repository/YaUserRepository.java
@@ -1,8 +1,8 @@
-package site.goldenticket.domain.yauser.repository;
+package site.goldenticket.dummy.yauser.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import site.goldenticket.domain.yauser.model.YaUser;
+import site.goldenticket.dummy.yauser.model.YaUser;
 
 @Repository
 public interface YaUserRepository extends JpaRepository<YaUser, Long> {
