+++ src/main/java/site/goldenticket/dummy/yauser/service/YaUserService.java
@@ -1,13 +1,15 @@
-package site.goldenticket.domain.yauser.service;
+package site.goldenticket.dummy.yauser.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import site.goldenticket.domain.yauser.repository.YaUserRepository;
+import site.goldenticket.dummy.yauser.repository.YaUserRepository;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
-@Slf4j
 public class YaUserService {
+
     private final YaUserRepository yaUserRepository;
+
 }
